requirements.txt
docker==6.0.1
dns-lexicon==3.11.7
ovh==1.0.0
certbot==2.1.0
cryptography==38.0.4
PyYAML==5.4.1
typed-argument-parser==1.7.2
